+++ dev-board/rf-messaging-module/data-link-layer.cpp
@@ -1,0 +1,29 @@
+//File Name: data-link-layer-layer.cpp
+//Description: This  handles the actual sending of messages over RF
+
+//Include Files
+#include "data-link-layer.h"
+
+//Private Constants
+
+//Private Variables
+
+//Public Constants
+
+//Public Variables
+
+//Private Function Prototypes
+
+//Public Function Prototypes
+
+//Private Function Definitions
+
+//Public Function Definitions
+RF_TRANSPORT_LAYER::RF_TRANSPORT_LAYER(USART_TypeDef* uart_instance):
+uartInstance{uart_instance}
+{}
+
+void sendMessage(uint8_t* data, uint16_t data_length)
+{
+
+}
