+++ dev-board/rf-messaging-module/tranport-layer.cpp
@@ -1,8 +1,8 @@
-//File Name: message-layer.cpp
+//File Name: transport-layer.cpp
 //Description: This implements the overall transmission and construction of an RF packet.
 
 //Include Files
-#include "message-layer.h"
+#include "tranport-layer.h"
 
 //Private Constants
 /*This is an example description for constants and variables. Delete this if it is unused here and copy and paste it to where it is needed. */
